requirements.txt
aiohttp==3.8.5
aio-pika==8.3.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3
pika==1.3.2
requests==2.32.3